+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/blocks/BlockCageTrap.java
@@ -47,7 +47,7 @@ public class BlockCageTrap extends BlockOwnable {
 		if(!deactivated){
 			return null;
 		}else{
-			return AxisAlignedBB.getBoundingBox((double)par2 + this.minX, (double)par3 + this.minY, (double)par4 + this.minZ, (double)par2 + this.maxX, (double)par3 + this.maxY, (double)par4 + this.maxZ);
+			return AxisAlignedBB.getBoundingBox(par2 + this.minX, par3 + this.minY, par4 + this.minZ, par2 + this.maxX, par3 + this.maxY, par4 + this.maxZ);
 		}
 	}
