+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/blocks/mines/BlockClaymore.java
@@ -66,7 +66,7 @@ public class BlockClaymore extends BlockContainer implements IExplosive {
 	}
 	
     public void onBlockPlacedBy(World par1World, int par2, int par3, int par4, EntityLivingBase par5EntityLivingBase, ItemStack par6ItemStack){
-        int l = MathHelper.floor_double((double)(par5EntityLivingBase.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+        int l = MathHelper.floor_double(par5EntityLivingBase.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 
         if(l == 0){
             par1World.setBlockMetadataWithNotify(par2, par3, par4, 1, 2);    
@@ -134,7 +134,7 @@ public class BlockClaymore extends BlockContainer implements IExplosive {
 	public void explode(World world, int par2, int par3, int par4) {
 		if(!world.isRemote){
 			BlockUtils.destroyBlock(world, par2, par3, par4, false);
-			world.createExplosion((Entity) null, (double) par2, (double) par3, (double) par4, 3.5F, true);
+			world.createExplosion((Entity) null, par2, par3, par4, 3.5F, true);
 		}
 	}
