+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/items/ItemModifiedBucket.java
@@ -137,10 +137,10 @@ public class ItemModifiedBucket extends Item {
                 return false;
             }else{
                 if (par1World.provider.isHellWorld && this.isFull == Blocks.flowing_water){
-                    par1World.playSoundEffect((double)((float)par2 + 0.5F), (double)((float)par3 + 0.5F), (double)((float)par4 + 0.5F), "random.fizz", 0.5F, 2.6F + (par1World.rand.nextFloat() - par1World.rand.nextFloat()) * 0.8F);
+                    par1World.playSoundEffect(par2 + 0.5F, par3 + 0.5F, par4 + 0.5F, "random.fizz", 0.5F, 2.6F + (par1World.rand.nextFloat() - par1World.rand.nextFloat()) * 0.8F);
 
                     for (int l = 0; l < 8; ++l){
-                        par1World.spawnParticle("largesmoke", (double)par2 + Math.random(), (double)par3 + Math.random(), (double)par4 + Math.random(), 0.0D, 0.0D, 0.0D);
+                        par1World.spawnParticle("largesmoke", par2 + Math.random(), par3 + Math.random(), par4 + Math.random(), 0.0D, 0.0D, 0.0D);
                     }
                 }else{
                     if (!par1World.isRemote && flag && !material.isLiquid()){
